Move selectedIds reducer up to avoid registration
