+++ packages/ra-core/src/reducer/admin/resource/index.spec.ts
@@ -28,7 +28,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'posts' },
                     },
@@ -42,7 +41,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'comments' },
                     },
@@ -66,7 +64,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'posts' },
             },
@@ -80,7 +77,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'comments' },
             },
@@ -94,7 +90,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'users', options: 'foo' },
             },
@@ -115,7 +110,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'posts' },
                     },
@@ -129,7 +123,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'comments' },
                     },
@@ -150,7 +143,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'posts' },
             },
@@ -171,7 +163,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'posts' },
                     },
@@ -185,7 +176,6 @@ describe('Resources Reducer', () => {
                                 sort: null,
                             },
                             expanded: [],
-                            selectedIds: [],
                         },
                         props: { name: 'comments' },
                     },
@@ -214,7 +204,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'posts' },
             },
@@ -228,7 +217,6 @@ describe('Resources Reducer', () => {
                         sort: null,
                     },
                     expanded: [],
-                    selectedIds: [],
                 },
                 props: { name: 'comments' },
             },
