+++ packages/ra-test/src/TestContext.spec.tsx
@@ -11,6 +11,7 @@ const primedStore = {
         notifications: [],
         resources: {},
         ui: {},
+        selectedIds: {},
     },
 };
